lre: Initial Flyway script

+++ src/main/resources/db/migration/V1__init-database.sql
@@ -2,7 +2,7 @@ drop table if exists beer cascade;
 drop table if exists customer cascade;
 
 create table beer (
-    beer_style          tinyint check (beer_style between 0 and 9),
+    beer_style          smallint check (beer_style between 0 and 9),
     price               numeric(38,2),
     quantity_on_hand    integer,
     version             integer,
@@ -11,7 +11,8 @@ create table beer (
     id                  varchar(36) not null,
     beer_name           varchar(255),
     upc                 varchar(255),
-    primary key (id));
+    primary key (id)
+);
 
 create table customer (
     version             integer,
@@ -19,4 +20,5 @@ create table customer (
     update_date         timestamp(6),
     id                  varchar(36) not null,
     name                varchar(255),
-    primary key (id));
+    primary key (id)
+ );
